Resolve VNC-500 "Bugfix/ sctp heartbeat length field underflow causes stack buffer overflow"

Closes VNC-500

See merge request kasm-technologies/internal/KasmVNC!275

+++ common/network/webudp/WuStun.h
@@ -53,5 +53,5 @@ struct StunPacket {
 
 bool ParseStun(const uint8_t* src, int32_t len, StunPacket* packet);
 
-int32_t SerializeStunPacket(const StunPacket* packet, const uint8_t* password,
+size_t SerializeStunPacket(const StunPacket* packet, const uint8_t* password,
                             int32_t passwordLen, uint8_t* dest, int32_t len);
